amended specification of dtype to follow SciPy recommendation

+++ python_examples/hit_and_miss.py
@@ -10,7 +10,7 @@ import numpy as np
 print('hit_and_miss')
 np.random.seed()
 
-r_0 = np.array([1.0,2.0,3.0],dtype='f8')
+r_0 = np.array([1.0,2.0,3.0],dtype=np.float_)
 v_0 = np.prod ( r_0 )
 
 tau_hit  = 0
