+++ python_examples/test_pot_linear.py
@@ -7,7 +7,7 @@ def random_positions(n):
     from maths_module import random_vector
     import sys
     
-    r = np.zeros((n,3),dtype='f8')
+    r = np.zeros((n,3),dtype=np.float_)
     # molecule 0 is always at the origin, now place the others randomly
     for i in range(1,r.shape[0]):
         for pos_try in range(npos):
@@ -29,7 +29,7 @@ def random_orientations(n):
     """Returns n random 3-d vectors in a numpy array (n,3)."""
     import numpy as np
     from maths_module import random_vector
-    e = np.zeros((n,3),dtype='f8')
+    e = np.zeros((n,3),dtype=np.float_)
     for i in range(e.shape[0]):
         e[i,:] = random_vector()
     return e
@@ -127,8 +127,8 @@ for i_xyz, f_exact in np.ndenumerate(f):
     print ( "{:5d}{:>10}{:15.6f}{:15.6f}{:15.4e}".format(i,cf[xyz],f_exact,fnum,f_exact-fnum) )
 
     t_exact = t[i_xyz] # Exact torque
-    axis = np.zeros(3,dtype='f8')
-    esave = np.zeros(3,dtype='f8')
+    axis = np.zeros(3,dtype=np.float_)
+    esave = np.zeros(3,dtype=np.float_)
     axis[xyz] = 1.0 # Pick axis
     esave[:] = e[i,:] # Save orientation vector (copy, not view)
     e[i,:] = rotate_vector ( delta, axis, esave ) # Rotate
